requirements.txt
aiogram
googletrans
Pillow
SQLAlchemy
